% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{catch_error_json}
\alias{catch_error_json}
\title{Helper function which catches (404) errors, when requesting a JSON from a remote source}
\usage{
catch_error_json(txt, quietly = TRUE)
}
\arguments{
\item{txt}{String: URL to the JSON file}

\item{quietly}{If FALSE, a warning message is returned, whenever an error is encountered. Default: TRUE}
}
\value{
A list with the query results or a list with NAs for each variable that was requested
}
\description{
Helper function which catches (404) errors, when requesting a JSON from a remote source
}
