% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-service.R
\name{coocsim}
\alias{coocsim}
\title{Lists words that are similar to the input word}
\usage{
coocsim(word, corpus = "deu_news_2012_1M", min_sim = 0.25,
  max_num = 20)
}
\arguments{
\item{word}{A string or vector of strings containing words, for which the information should be queried}

\item{corpus}{Name of the corpus to query. Default is the German deu_news_2012_1M}

\item{min_sim}{Minimal similarity of the possible candidate words to the queried word}

\item{max_num}{integer. Maximal number of retrieved similar words.}
}
\value{
A tibble with words similar to the queried_word
}
\description{
The results are based on the sentence and neighbour cooccurrences profiles of the words in descending order based on similarity value.
}
\examples{
coocsim("mein")

coocsim(c("mein", "ist"), min_sim=0.3, max_num=4)
}
