% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{list_available_corpora}
\alias{list_available_corpora}
\title{List the available corpora of Leipzig Corpora Collection}
\usage{
list_available_corpora(name_only = FALSE)
}
\arguments{
\item{name_only}{logical. If TRUE, only the names of the corpora will be return.}
}
\value{
Either a tibble with all the corpora and their description or a character vector with all corpora names
}
\description{
List the available corpora of Leipzig Corpora Collection
}
\examples{
list_available_corpora()
list_available_corpora(TRUE)
}
