% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-service.R
\name{prefixword}
\alias{prefixword}
\title{Query words starting with a specific letter combination}
\usage{
prefixword(prefix, max_num = 10, min_freq = 2,
  corpus = "deu_news_2012_1M", force = FALSE)
}
\arguments{
\item{prefix}{A string or vector of strings containing the prefixes, for which words should be queried.
The prefixes must be at least 4 letters long.}

\item{max_num}{integer. Maximal number of retrieved words. Should be between 1 and 1000.}

\item{min_freq}{interger. Minimum frequency of retrieved words. Default is 2.}

\item{corpus}{string. Name of the corpus to query. Default is the German deu_news_2012_1M}

\item{force}{logical. Set to TRUE, if you want to query more than 1000 words}
}
\value{
A tibble with words starting with the prefix and their Frequency.
}
\description{
lists words with a specific prefix with a minimum length of 4 characters.
Can be used for autocompletion functionality or similar.
}
\examples{
prefixword("fort")
}
