% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-service.R
\name{query_prefix}
\alias{query_prefix}
\title{Wrapper around catch_error_json to create the right output for querys with errors}
\usage{
query_prefix(url)
}
\arguments{
\item{url}{URL for the API request}
}
\value{
list with the query results or a list with NAs
}
\description{
This function is used when querying prefixes and takes the result of the error catching json function as returns either results
if the request was successful or a list with the same dimensions as a successful request but with NAs
}
