% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-service.R
\name{query_similarity}
\alias{query_similarity}
\title{Wrapper around catch_error_json to create the right output for querys
to the similarity services with errors}
\usage{
query_similarity(url)
}
\arguments{
\item{url}{URL for the API request}
}
\value{
list with the query results or a list with NAs
}
\description{
This function is used when querying to the similarity service and takes the result of the error catching json function and returns either results
if the request was successful or a list with the same dimensions as a successful request but with NAs
}
