% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-service.R
\name{randomword}
\alias{randomword}
\title{Query a random selection of words with their frequency}
\usage{
randomword(n, corpus = "deu_news_2012_1M", force = FALSE)
}
\arguments{
\item{n}{integer: Number of words to query. Should be between 1 and 1000}

\item{corpus}{Name of the corpus to query. Default is the German deu_news_2012_1M}

\item{force}{logical. Set to TRUE, if you want to query more than 1000 words}
}
\value{
A tibble with random words, their id and their word frequency
}
\description{
lists randomly choosen words (of medium frequency) of a corpus.
Can be used to provide the user with a preselection of words.
}
\examples{
randomword(5)

}
