% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cooccurences.R
\name{retrieve_right_cooccurences}
\alias{retrieve_right_cooccurences}
\title{Retrieve right cooccurences for a specific word from Leipzig Wortschatz}
\usage{
retrieve_right_cooccurences(wordlist, limit = 20, corpus = "deu_news_2012")
}
\arguments{
\item{wordlist}{A character vector containing the words, for which the data is requested. Other types will be coerced into characters.}

\item{limit}{Number of most significant right cooccurrences. Must be larger than 2}

\item{corpus}{Name of the corpus to be used. At the moment only "deu_news_2012" possible}
}
\value{
data.frame with the requested right cooccurence frequencies
}
\description{
Retrieve right cooccurences for a specific word from Leipzig Wortschatz
}
\examples{
my_wordlist <- c("Hund", "Faultier", "Kruoz")
my_word_data <- retrieve_right_cooccurences(my_wordlist)
my_word_data
}
