% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_word_data_func.R
\name{retrieve_word_data}
\alias{retrieve_word_data}
\title{Retrieve frequency data from Leipzig Wortschatz}
\usage{
retrieve_word_data(wordlist, corpus = "deu_news_2012", as_dataframe = TRUE)
}
\arguments{
\item{wordlist}{A character vector containing the words, for which the data is requested}

\item{corpus}{Name of the corpus to be used. At the moment only "deu_news_2012" possible}

\item{as_dataframe}{logical; if True (default) a data.frame is returned, else a list}
}
\value{
Either data.frame or list with the requested frequency data
}
\description{
Retrieve frequency data from Leipzig Wortschatz
}
\examples{
my_wordlist <- c("Hund", "Faultier", "Kruoz")
my_word_data <- retrieve_word_data(my_wordlist)
my_wort_data
}
