% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-service.R
\name{word}
\alias{word}
\title{Queries a lists of information about a word in a corpus.}
\usage{
word(word, corpus = "deu_news_2012_1M", ...)
}
\arguments{
\item{word}{A string or vector of strings containing words, for which the information should be queried}

\item{corpus}{Name of the corpus to query. Default is the German deu_news_2012_1M}

\item{...}{Additional arguments passed to catch_error_json. For now quietly = FALSE enables the display of warnings,
whenever a word has not entry in the corpus}
}
\value{
A tibble with the information about the queried words. That is Frequency, FrequencyClass and WordRank
}
\description{
Queries a lists of information about a word in a corpus.
}
\examples{

word(c("Hund", "Katze", "Maus"))

}
